MultimodalRAG/requirements_bak.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.6.3
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==45.0.7
cyclopts==3.24.0
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.117.1
fastjsonschema==2.21.2
fastmcp==2.12.2
ffmpy==0.6.1
filelock==3.13.1
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gradio==5.46.1
gradio_client==1.13.1
greenlet==3.2.4
groovy==0.1.2
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
labelImg==1.8.6
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.29
lark==1.2.2
lazy-object-proxy==1.12.0
loguru==0.7.3
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.14.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.1.2
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pymongo==4.14.1
pyparsing==3.2.5
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
segment-anything==1.0
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
sympy==1.13.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
typer==0.17.4
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
widgetsnbextension==4.0.14
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
